Moved login tests to platform.
